pip_requirements.txt
absl-py==1.4.0
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.94.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
jax==0.4.8
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.12.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.6
Pygments==2.14.0
pymongo==4.3.3
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.0.post1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
yarg==0.1.9
